Scaffold creation of Game resource with name attribute
A 'Game' is just a container for a number of points on a grid that will
change based on the rules of Conway's game of life. The points will be a
separate model, which will later facilitate introducing more complex
rules and different types of "creatures".

Because the 'Game' model just serves as the owner of many 'Life' models,
it just needs a name. The foreign key will be on the 'Life' model.