Add API games controller to return JSON game data
Our Flux JS app is going to be making requests of our API to get and
save data about games, as well as using the data bootstrapped into the
original page. I like having a separate set of routes for the restful
JSON API rather than having the single games controller do double-duty
of serving JSON and static HTML.  I think in the long run this will be
easier to maintain, and we can still reuse the JBuilder templates
between the two controllers.