Add LifeApp to contain EcosystemGrid for a given game
Instead of rendering a 'show' page for each game and rendering the
EcosystemGrid on the 'show' page, this will be a truly one page app.

The LifeApp will have access to all games via the GamesStores, and will
render a GamesIndex sidebar that can change the 'currentGame' in the
GamesStores.