Fix syntax to use instance of Dispatcher
The Flux.Dispatcher is a constructor, and we need to use an instance of
it to actually call methods. Using 'extend' or 'merge' didn't seem to
work to create a constructor that inherits from Flux.Dispatch, so I
hacked a set of helper methods that close over the instance of
Dispatcher, exposing a modified and limited set of methods.