Add GridLoadingScreen component
The only reason to make this it's own component was for the sake of
organization, following the single responsability principle. The
LifeApp's only responsiblity is to hold all the pieces of the page and
pass them the right data - it shouldn't contain the logic or
presentation for any parts of the screen within itself.