Add cell count for each game listed in GameIndex
To further explore the way that Flux keeps everything in sync, I added a
count of cells to the listings in the GameIndex. It's neat to see them
live-update when you save a game with a new set of cells.

There is a bug where the cell count doesn't update when you only remove
cells. Looking into this next.