Fix bug where removed cells were not deleted on save
The server was returning the wrong data after 'save' because, even
though we did delete the old 'live' cells and make completely new ones,
the 'game' object was not reloaded and did not see the changes to it's
associated 'life' models.

We were also using the lodash deep 'merge' to combine the attributes
returned by the server with the previous attributes, when really we can
safely just overwrite the old attributes, trusting what the server
returns (now that it is correct).