Update EcosystemGrid based on currentGame in stores
The ecosystemGrid gets a game as a prop from the lifeApp. To get it to
re-render it's tiles, (since there is no automatic state change when a
new game is passed in as a prop), we added a hook into the
componentWillReceiveProps lifecycle method to update the state. I don't
like the use of props to generate state, and I'm going to brainstorm a
different approach. The EcosystemGrid is watching the stores - maybe
instead of having a game passed in, it should just check the stores for
the latest 'currentGame' on every render?