Add 'new' game button; fixes #3
What is neat is that very little of the logic for this new action was
added to the React component - just a button, and then the action
updating the stories caused the right updates in all the other
components.

I'm still figuring out which parts should be in the 'action creators'
and which can go in helper methods like the 'game_web_api_utils'.