Improve styles and add scroll indicating shadows to GameIndex
Now when there is room to scroll to the bottom of the GameIndex, there
will be an inner shadow on the bottom edge, and likewise at the top when
there is room to scroll to the top. Having both shadows appear when the
scroll bar is in the middle was tricky - I didn't notice at first that I
couldn't apply styles cumulatively with Sass and placeholders, so the
second placeholder was overriding the first.

It still doesn't look right, but I think this makes it more usable.