Break out GridTile component from EcosystemGrid
For the moment, this takes a tiny bit of the logic and a lot of the
styles out of the concerns of the EcosystemGrid, but most of the logic
for the Game of Life is still held within the grid.

That is because the GridTile will soon hold more logic for rendering
different types of tiles.