Add 'Reset' button
The 'reset' button gets the 'currentGame' from the stores without saving
any of the local changes and overwrites the local changes with the
version in the stores.

Next I will add a mini-view of the saved state of the game, so that it
can be compared with the edited state. There should be some text and a
colored overlay that change to indicated whether your changes have been
saved to the stores/server or if they are out of sync.