Add preview grid to top right corner
It's nice to see the original/saved version of the current game, and now
that is available in a tiny preview in the top right corner. Generating
this is so easy, because it can grab data directly from the Stores.

I think a good refactor would be to create a controlled Grid component
and use that in both GridPreview and EcosystemGrid.